fix: code moved from utils to api
